feat: implement factory for collector initialization
